Makefiles: add rules to make the .depend files if they were deleted
For now, the .depend file is only rebuilt if not existing, it is not rebuilt when needed.
And it is not built when running make clean.
The target 'depend' can still be used to force rebuilding the .depend file.

To goal is to remove the .depend files from the repository because them may depend on the configuration.